moves isolated AGSMapView+ExtentForCGRect to caller file
