+++ data-collection/data-collection/View Controllers/Map View Controller/MapViewController+LocationSelection.swift
@@ -249,3 +249,35 @@ private extension AGSMapView {
         return screen(toLocation: bounds.center)
     }
 }
+
+// MARK: Extent from CGRect
+
+private extension AGSMapView {
+    
+    /// Convert a `CGRect` to an `AGSPolygon` in relation to the `AGSMapView`'s bounds.
+    ///
+    /// - Parameter rect: The rect in point space that is to be converted to a spatial rectangular polygon.
+    ///
+    /// - Returns: The newly created geometry.
+    ///
+    /// - Throws: An error if the `fromRect` parameter cannot be contained by the map view's bounds.
+    
+    func convertExtent(fromRect rect: CGRect) throws -> AGSGeometry {
+        
+        guard bounds.contains(rect) else {
+            fatalError("The CGRect \(rect) provided is outside of the map view's bounds \(bounds).")
+        }
+        
+        let nw = rect.origin
+        let ne = CGPoint(x: rect.maxX, y: rect.minY)
+        let se = CGPoint(x: rect.maxX, y: rect.maxY)
+        let sw = CGPoint(x: rect.minX, y: rect.maxY)
+
+        let agsNW = screen(toLocation: nw)
+        let agsNE = screen(toLocation: ne)
+        let agsSE = screen(toLocation: se)
+        let agsSW = screen(toLocation: sw)
+
+        return AGSPolygon(points: [agsNW, agsNE, agsSE, agsSW])
+    }
+}
